% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scRefBool.R
\name{run_scRefBool}
\alias{run_scRefBool}
\title{scRefBool: quantifies different levels of gene expression including low, medium and high}
\usage{
run_scRefBool(
  data,
  metadata,
  org = "human",
  dir,
  file.name,
  precBack = NA,
  discretize = TRUE,
  sig.frames = TRUE,
  ncores = detectCores() - 2,
  fixseed = 1234
)
}
\arguments{
\item{data}{single cell expression matrix}

\item{metadata}{corresponding metadata (two columns: cell names, cell type)}

\item{org}{organism (human or mouse)}

\item{dir}{directory used to save the results}

\item{file.name}{prefix for output files and folder}

\item{precBack}{precompiled background path to ScRbRef}

\item{discretize}{gene quantification in three levels of cell types}

\item{sig.frames}{construction of significant genes data frame}

\item{ncores}{number of cores to use to run the code}

\item{fixseed}{seed used to discretize cell types (reproducbility)}
}
\value{
NULL, all results are accessible in the dir specified
}
\description{
scRefBool: quantifies different levels of gene expression including low, medium and high
}
\author{
Céline Barlier, Kartikeya Singh, Sascha Jung
}
